+++ src/main/java/com/weshare/api/v1/controller/auth/advice/AuthenticationExceptionHandler.java
@@ -8,16 +8,14 @@ import com.weshare.api.v1.token.exception.TokenTimeOutException;
 import com.weshare.api.v1.common.Response;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.support.DefaultMessageSourceResolvable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.validation.BindException;
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.context.request.WebRequest;
+
 
 import static com.weshare.api.v1.controller.auth.AuthErrorCode.*;
 
@@ -29,32 +27,10 @@ public class AuthenticationExceptionHandler {
     private final Response response;
 
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity illegalArgumentExceptionHandler (IllegalArgumentException e){
-        log.error("[exceptionHandler] ex", e);
-        return response.fail(BAD_REQUEST_ERROR.getCode(), HttpStatus.BAD_REQUEST, e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(OAuthApiException.class)
     public ResponseEntity oAuthApiExceptionHandler (OAuthApiException e){
         log.error("[exceptionHandler] ex", e);
         return response.fail(AUTH_LOGIN_BAD_REQUEST.getCode(), HttpStatus.BAD_REQUEST, e.getMessage());
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(MethodArgumentNotValidException.class)
-    public ResponseEntity methodArgumentNotValidExceptionHandler (MethodArgumentNotValidException  e){
-        log.error("[exceptionHandler] ex", e);
-        String[] errorMessages = getDefaultErrorMessage(e);
-        return response.fail(PARAMETER_BAD_REQUEST_ERROR.getCode(), HttpStatus.BAD_REQUEST, errorMessages);
-    }
-
-    private String[] getDefaultErrorMessage(BindException e) {
-        return e.getBindingResult().getFieldErrors()
-                .stream()
-                .map(DefaultMessageSourceResolvable::getDefaultMessage)
-                .toArray(String[]::new);
     }
 
     @ResponseStatus(HttpStatus.UNAUTHORIZED)
@@ -90,13 +66,5 @@ public class AuthenticationExceptionHandler {
     public ResponseEntity EmailDuplicateExceptionHandler (EmailDuplicateException e) {
         log.error("[exceptionHandler] ex", e);
         return response.fail(EMAIL_DUPLICATE_ERROR.getCode(), HttpStatus.CONFLICT, e.getMessage());
-    }
-
-
-    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
-    @ExceptionHandler
-    public ResponseEntity exHandler(Exception e) {
-        log.error("[exceptionHandler] ex", e);
-        return response.fail(SERVER_ERROR.getCode(), HttpStatus.INTERNAL_SERVER_ERROR, "[server error] " + getDefaultErrorMessage((BindException) e));
     }
 }
